requirements.txt
pandas
numpy
beautifulsoup4
google-play-scraper
nltk
vaderSentiment
textblob
scikit-learn
tensorflow
transformers
matplotlib
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
spacy
wordcloud
matplotlib
seaborn
yake
rake-nltk
